Fix: 폴더구조 개선, 컴포넌트 세분화
