+++ src/app/login/api/login.ts
@@ -1,0 +1,20 @@
+import { LoginCredentials, LoginResponse } from "../types";
+
+export async function loginUser(
+  credentials: LoginCredentials
+): Promise<LoginResponse> {
+  const response = await fetch("/api/auth/login", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(credentials),
+  });
+
+  if (!response.ok) {
+    const errorData = await response.json();
+    throw new Error(errorData.message || "Login failed");
+  }
+
+  return response.json();
+}
