+++ src/app/posts/types/index.ts
@@ -1,0 +1,14 @@
+export interface Post {
+  id: string;
+  title: string;
+  content: string;
+  images: string[];
+}
+
+export interface CommentProps {
+  username: string;
+  content: string;
+  timestamp: string;
+  depth: number;
+  replies?: CommentProps[];
+}
