+++ src/app/profile/types/index.ts
@@ -1,0 +1,9 @@
+export interface ProfileData {
+  nickname: string;
+  name: string;
+  email: string;
+  birthDate: string;
+  followers: number;
+  following: number;
+  imageUrl?: string;
+}
