+++ src/app/signup/api/signup.ts
@@ -1,0 +1,19 @@
+import type { RegisterUserData, RegisterResponse } from "../types";
+
+export async function registerUser(
+  userData: RegisterUserData
+): Promise<RegisterResponse> {
+  const response = await fetch("/api/auth/register", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(userData),
+  });
+
+  if (!response.ok) {
+    throw new Error("Registration failed");
+  }
+
+  return response.json();
+}
