+++ src/components/common/ui/ContentBox.tsx
